+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please visit https://github.com/shd101wyy/vscode-markdown-preview-enhanced/relea
 
 - Updated `mermaid` version to the latest `11.5.0`.
 - Updated `katex` version to the latest `0.16.21`.
+- Updated `prismjs` version to the latest `1.30.0`.
 
 ## [0.9.11] - 2024-11-08